Remove ~ that people keep putting in the system variable (#224)
For the record, `~0x20000000` clears the 29th bit of the result from cpuid, while `0x20000000` replaces the result with one with just the 29th bit set (OPENSSL_ia32cap.3). The 29th bit represents the Intel SHA Extensions.

As such, if the second one (without the ~) works, the actual solution has nothing to do with the 29th bit; it's one of the others.